appsrc video capture, tick-based pipeline option

+++ GStreamer/Source/GStreamer/Private/GstAppSrcImpl.h
@@ -9,5 +9,5 @@ class IGstAppSrc
   public:
 	virtual bool Connect(class IGstPipeline *Pipeline, const char *ElementName) = 0;
 	virtual void Disconnect() = 0;
-	virtual void PushTexture(const uint32_t *TextureData, size_t TextureSize) = 0;
+	virtual void PushTexture(const uint8_t *TextureData, size_t TextureSize) = 0;
 };
